feat: Remove deprecated return value on .js and .css methods
BREAKING CHANGE: In version 4 of podium it became possible to set multiple assets to a layout through its `.js()` and `.css()` methods. This did make it impossible to let these methods return a reasonable value, but to keep a backwards compatibility with version 3 of Podium, the first item passed in was returned for compatibility. This removes this compatibility with version 3 of Podium.

This resolves https://github.com/podium-lib/issues/issues/23