feat: reexport types from podium/client
For TypeScript users sometimes the podium/client package (used only
for types) would be updated to a newer version than what was used
in the latest podium/layout package. This would cause unexpected
type incompatibilities. From this version on, export the types
from podium/client so TypeScript users can remove the direct
dependency on podium/client, ensuring the types are always in sync
with podium/layout.